third_party/blink/web_tests/fast/borders/border-width-percent-expected.txt

This tests the border width property with percent values for HTML elements.

PASS elementBorderWidth('img', '10%') is 10
PASS elementBorderWidth('img', '-10%') is 0
PASS elementBorderWidth('img', ' +10%') is 10
PASS elementBorderWidth('img', 0, 'border-width: 10%') is 0
PASS elementBorderWidth('img', 0, 'border-width: -10%') is 0
PASS elementBorderWidth('object', '10%') is 10
PASS elementBorderWidth('object', '-10%') is 0
PASS elementBorderWidth('object', ' +10%') is 10
PASS elementBorderWidth('object', 0, 'border-width: 10%') is 0
PASS elementBorderWidth('object', 0, 'border-width: -10%') is 0
PASS inputElementBorderWidth('image', '10%') is 10
PASS inputElementBorderWidth('image', '-10%') is 0
PASS inputElementBorderWidth('image', ' +10%') is 10
PASS inputElementBorderWidth('image', 0, 'border-width: 10%') is 0
PASS inputElementBorderWidth('image', 0, 'border-width: -10%') is 0
PASS successfullyParsed is true

TEST COMPLETE
